+++ client/web/apis/trending.ts
@@ -1,5 +1,5 @@
 import axiosInstance from '@/lib/axios.config';
-import { TMDB_API } from '../constants';
+import { TMDB_API } from '@/utils/constants';
 
 export const fetchTrending = (mediaType: 'all' | 'movie' | 'tv' | 'person', timeWindow: 'day' | 'week') => {
   return axiosInstance.get(TMDB_API.TRENDING(mediaType, timeWindow));
