+++ client/web/package.json
@@ -12,7 +12,6 @@
     "@clerk/nextjs": "^6.5.1",
     "@radix-ui/react-slot": "^1.1.0",
     "@radix-ui/react-tooltip": "^1.1.4",
-    "@tailwindcss/line-clamp": "^0.4.4",
     "axios": "^1.7.9",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
