+++ client/web/tailwind.config.ts
@@ -1,6 +1,5 @@
 import type { Config } from 'tailwindcss';
 import tailwindcssAnimate from 'tailwindcss-animate';
-import lineClamp from '@tailwindcss/line-clamp';
 
 export default {
   darkMode: ['class'],
@@ -65,5 +64,5 @@ export default {
       },
     },
   },
-  plugins: [tailwindcssAnimate, lineClamp],
+  plugins: [tailwindcssAnimate],
 } satisfies Config;
